refactor(admin): reorganize settings directory structure
## 変更内容

admin/settings ディレクトリを削除し、各設定画面のコンポーネントを適切な場所に移動しました。

### 削除
- `app/[locale]/admin/settings/page.tsx` - リダイレクト専用の空ページ

### 移動

#### フォームコンポーネント → 各ページの _components ディレクトリへ
- `settings/_components/general-settings-form.tsx` → `general/_components/`
- `settings/_components/auth-providers-form.tsx` → `gatekeeper/_components/`
- `settings/_components/resource-limits-form.tsx` → `resource-limits/_components/`
- `settings/_components/expiration-archive-form.tsx` → `expiration-archive/_components/`

#### 共有ファイル → admin 直下の共有ディレクトリへ
- `settings/_actions/system-settings.ts` → `admin/_actions/`
- `settings/_actions/__tests__/system-settings.test.ts` → `admin/_actions/__tests__/`
- `settings/_lib/form-options.ts` → `admin/_lib/`

### インポートパスの更新
すべてのインポートパスを新しいディレクトリ構造に合わせて更新しました。

## メリット

- 各設定ページが独自のコンポーネントを持つ明確な構造
- 共通ロジック（actions と form options）は admin 配下で共有
- 不要なリダイレクトページを削除
- より保守しやすいディレクトリ構造

## 検証

- [x] 型チェック: `pnpm type-check` 成功
- [x] テスト: `pnpm test app/[locale]/admin/_actions/__tests__/system-settings.test.ts` 成功
- [x] Lint: `pnpm check` 成功
