+++ app/[locale]/admin/general/_components/general-settings-form.tsx
@@ -32,11 +32,11 @@ import {
 } from "@/components/ui/select";
 import type { SystemSettings } from "@/lib/schemas/system-settings";
 import { getErrorMessage } from "@/lib/utils/error-message";
-import { updateSystemSettingsAction } from "../_actions/system-settings";
+import { updateSystemSettingsAction } from "../../_actions/system-settings";
 import {
   systemSettingsFormOpts,
   systemSettingsFormSchema,
-} from "../_lib/form-options";
+} from "../../_lib/form-options";
 
 interface Props {
   initialSettings?: SystemSettings;
