+++ app/[locale]/admin/general/page.tsx
@@ -1,6 +1,6 @@
 import { getTranslations, setRequestLocale } from "next-intl/server";
 import { getSystemSettings } from "@/lib/data/system-settings";
-import { GeneralSettingsForm } from "../settings/_components/general-settings-form";
+import { GeneralSettingsForm } from "./_components/general-settings-form";
 
 export const dynamic = "force-dynamic";
