+++ app/[locale]/admin/resource-limits/_components/resource-limits-form.tsx
@@ -25,11 +25,11 @@ import {
 import { Input } from "@/components/ui/input";
 import type { SystemSettings } from "@/lib/schemas/system-settings";
 import { getErrorMessage } from "@/lib/utils/error-message";
-import { updateSystemSettingsAction } from "../_actions/system-settings";
+import { updateSystemSettingsAction } from "../../_actions/system-settings";
 import {
   systemSettingsFormOpts,
   systemSettingsFormSchema,
-} from "../_lib/form-options";
+} from "../../_lib/form-options";
 
 interface Props {
   initialSettings?: SystemSettings;
