--- app/[locale]/admin/settings/page.tsx
@@ -1,14 +1,0 @@
-import { redirect } from "next/navigation";
-import { setRequestLocale } from "next-intl/server";
-
-export const dynamic = "force-dynamic";
-
-export default async function AdminSettingsPage({
-  params,
-}: PageProps<"/[locale]/admin/settings">) {
-  const { locale } = await params;
-  setRequestLocale(locale);
-
-  // Redirect to the general settings page
-  redirect(`/${locale}/admin/general`);
-}
